[stable @ 65a3718b2af2843d50ca9657ec123689e458524b] Windows command line parsing: go with different strategy regarding blackslashes inside quoted parameters ; more fixes from Henry Nestler
git-svn-id: svn://svn.code.sf.net/p/colinux/code/branches/stable@154 688f6386-cf32-0410-9ead-a596f0484e40

+++ src/colinux/os/linux/user/pipe.c
@@ -217,7 +217,7 @@ co_rc_t co_os_pipe_get_colinux_pipe_path(co_id_t instance, char *out_path, int s
 	if (home == NULL)
 		home = "/tmp";
 
-	snprintf(out_path, size, "%s/.colinux.%d.pipe", home, instance);
+	snprintf(out_path, size, "%s/.colinux.%d.pipe", home, (int) instance);
 
 	return CO_RC(OK);
 }
